Added sum diff percent10h

+++ src/main/resources/db/migration/20181123204623__Create_table_stats.sql
@@ -3,8 +3,8 @@ CREATE SEQUENCE seq_statistic_id START 1 INCREMENT BY 300;
 CREATE TABLE statistic (
     id BIGINT NOT NULL DEFAULT nextval('seq_statistic_id'),
     created_at BIGINT,
-    all_cryptos NUMERIC(20, 8),
-    valid_cryptos NUMERIC(20, 8),
+    all_cryptos BIGINT,
+    valid_cryptos BIGINT,
     probability NUMERIC(20, 8),
 
     CONSTRAINT pk_statistic_id PRIMARY KEY (id)
